% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_to_query_string.R
\name{arg_to_query_string}
\alias{arg_to_query_string}
\title{Build API query string}
\usage{
arg_to_query_string(fld, arg, quotes = TRUE)
}
\arguments{
\item{fld}{\itemize{
\item the name of the field to be queried
}}

\item{arg}{\itemize{
\item the value(s) to passed arg_to_query_string.R
}}

\item{quotes}{\itemize{
\item boolean - should the values be wrapped in quotes?
}}
}
\value{
query string
}
\description{
Build the api query string
}
\details{
This function accepts two argulements - a string which representes
the field name and a string, or vector of strings that will be
used to construct the api query string.  Single fields will result
in a query string of the form "\if{html}{\out{<FLD>}}=\if{html}{\out{<ARG>}}", multiple arguments
will produce a string of the form "\if{html}{\out{<FLD>}}+IN+(ARG1, ARG2,...)".
Some arguments may need to be wrapped in quotes in the final
string, 'quotes=TRUE' with ensure that is the case.
}
\author{
R. Adam Cottrill
}
