% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_waterbody.R
\name{build_name_ends_with}
\alias{build_name_ends_with}
\title{Build waterbody name ends with}
\usage{
build_name_ends_with(name_ends_with)
}
\arguments{
\item{name_ends_with}{\itemize{
\item string to match the end of waterbody name fields to
}}
}
\value{
partial api query string
}
\description{
Build waterbody name ends with
}
\details{
A helper function used by fetch_waterbody that will take a string
and return the api query string that will match the end of the
unofficial name, offical name, equivalent french name, and
official alternative name.
}
\author{
R. Adam Cottrill
}
