% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_waterbody.R
\name{build_name_like}
\alias{build_name_like}
\title{Build waterbody name starts with}
\usage{
build_name_like(name_like)
}
\arguments{
\item{name_like}{\itemize{
\item string to match part of waterbody name fields to
}}
}
\value{
partial api query string
}
\description{
Build waterbody name contains
}
\details{
A helper function used by fetch_waterbody that will take a string
and return the api query string that will match any waterbodies
where the unofficial name, offical name, equivalent french name, and
official alternative name contain the string.
}
\author{
R. Adam Cottrill
}
