% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_conservation_authority.R
\name{fetch_conservation_authority}
\alias{fetch_conservation_authority}
\title{Fetch Conservation Authority data from Geohub}
\usage{
fetch_conservation_authority(name_like = NULL, name_starts_with = NULL)
}
\arguments{
\item{name_like}{\itemize{
\item a string to match against part of the common
name field.  This can be a single string or character
vector of strings.
}}

\item{name_starts_with}{\itemize{
\item a string to match against the start of
the common name field. This can be a single string or
character vector of strings.
}}
}
\value{
a sf object containing the geometry or geometries of the
matching Conservation Authority object(s)
}
\description{
Fetch Conservation Authority objects from Geohub
}
\details{
This function will connect to the Geohub api and return the
matching records from the Conservation Authority
endpoint.  The geohub details page can be found here:
\url{https://geohub.lio.gov.on.ca/datasets/lio::conservation-authority-administrative-area/about}

This function currently supports filters for COMMON_NAME like or
COMMON_NAME starts with.  Please note that the geohub api is
case sensitive and that the Conservation Authority names are in 'title
case'. This means that "Saugeen" will match the 'Saugeen Conservation',
but 'SAUGEEN' and 'saugeen' will not.
}
\examples{

\donttest{
ca <- fetch_conservation_authority(name_like='Saugeen')
unique(ca$COMMON_NAME)

ca <- fetch_conservation_authority(name_like=c('Saugeen', 'Maitland'))
unique(ca$COMMON_NAME)

my_plot <- ggplot2::ggplot() +
 ggplot2::geom_sf(data = ca, ggplot2::aes(fill=COMMON_NAME))
print(my_plot)
}

}
\author{
R. Adam Cottrill
}
