% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_fmz.R
\name{fetch_fmz}
\alias{fetch_fmz}
\title{Fetch Fisheries Management Zones from Geohub}
\usage{
fetch_fmz(fmz_number)
}
\arguments{
\item{fmz_number}{\itemize{
\item a number or vector of numbers corresponding the
Fisheries Management Zone numbers to retrieve from geohub.
}}
}
\value{
a sf obect containing the geometry or geometries of the
provided fisheries management zone(s).
}
\description{
Fetch fmz objects from Geohub
}
\details{
This function will connect to the Geohub api and return the
matching records from the Ontario Fisheries Management Zone
endpoint.  The geohub details page can be fount here:
https://geohub.lio.gov.on.ca/datasets/lio::fisheries-management-zone/about

This function current supports filters for fisheries management
zone id - this can be a single integer, or a vector of integers
(e.g. c(4,5,6)).  Only values between 1 and 20 are allowed.
.
}
\author{
R. Adam Cottrill
}
