% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_geojson_as_sf.R
\name{fetch_geojson_as_sf}
\alias{fetch_geojson_as_sf}
\title{Request a resource from geohub}
\usage{
fetch_geojson_as_sf(url)
}
\arguments{
\item{url}{\itemize{
\item the url (including query parameters) to the requested
geohub resource.
}}
}
\value{
sf (simple feature) object
}
\description{
Request a geojson resource from geohub
}
\details{
This functions sends a request to the provide url and attempts to
covert the geojson repsonse into a sf (simple feature) object. If
the response is not successful, an error is thrown.  This funciton
is not expected to be used directly by user, but is called by most
of other functions in the OntarioGeohubR package.
}
\author{
R. Adam Cottrill
}
