% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_mnr_district.R
\name{fetch_mnr_district}
\alias{fetch_mnr_district}
\title{Fetch MNR Districts from Geohub}
\usage{
fetch_mnr_district(name_like = NULL, name_starts_with = NULL)
}
\arguments{
\item{name_like}{\itemize{
\item a string to match against part of the district
name field.  This can be a single string or character
vector of strings.
}}

\item{name_starts_with}{\itemize{
\item a string to match against the start of
the district name field. This can be a single string or
character vector of strings.
}}
}
\value{
a sf object containing the geometry or geometries of the
provided fisheries management zone(s).
}
\description{
Fetch MNR District objects from Geohub
}
\details{
This function will connect to the Geohub api and return the
matching records from the MNR Disticts endpoint.  The geohub
details page can be found here:
\url{https://geohub.lio.gov.on.ca/datasets/lio::mnr-district/about}

This function currently supports filters for DISTRICT_NAME like or
DISTRICT_NAME starts with.  Please note that the geohub api is
case sensitive and that the MNR districts are in 'title
case'. This means that "Aurora" will match the 'Aurora District',
but 'AURORA' and 'aurora' will not.
}
\examples{

\donttest{
dist <- fetch_mnr_district(name_like='Aurora')
unique(dist$DISTRICT_NAME)

dist <- fetch_mnr_district(name_like=c('Aurora', 'Owen Sound'))
unique(dist$DISTRICT_NAME)

my_plot <- ggplot2::ggplot() +
 ggplot2::geom_sf(data = dist)
print(my_plot)
}

}
\author{
R. Adam Cottrill
}
