% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_provincial_park.R
\name{fetch_provincial_park}
\alias{fetch_provincial_park}
\title{Fetch Provincial Park Data from  Geohub}
\usage{
fetch_provincial_park(name_like = NULL, name_starts_with = NULL)
}
\arguments{
\item{name_like}{\itemize{
\item a string to match against part of the common short
name field.  This can be a single string or character
vector of strings.
}}

\item{name_starts_with}{\itemize{
\item a string to match against the start of
the common short name field. This can be a single string or
character vector of strings.
}}
}
\value{
a sf object containing the geometry or geometries of the
matching provincal park(s).
}
\description{
Fetch Provincial Park objects from Geohub
}
\details{
This function will connect to the Geohub api and return the
matching records from the Ontario Provincial Park
endpoint.  The geohub details page can be found here:

\url{https://geohub.lio.gov.on.ca/datasets/lio::provincial-park-regulated/about}

This function current supports filters for COMMON_SHORT_NAME like
or COMMON_SHORT_NAME starts with.  Please note that the geohub
api is case sensitive and that the Provincial Parks are stored in
uppercase.  This function will automatically convert any filter
arguments to upper case before sending the request.
}
\examples{

\donttest{
ppark <- fetch_provincial_park(name_like='Algonquin')
unique(ppark$COMMON_SHORT_NAME)

ppark <- fetch_provincial_park(name_like=c('Algonquin', 'Inverhuron'))
unique(ppark$COMMON_SHORT_NAME)

my_plot <- ggplot2::ggplot() +
 ggplot2::geom_sf(data = ppark, ggplot2::aes(fill=COMMON_SHORT_NAME))
print(my_plot)
}

}
\author{
R. Adam Cottrill
}
