% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_township.R
\name{fetch_township}
\alias{fetch_township}
\title{Fetch Township Data from  Geohub}
\usage{
fetch_township(name_like = NULL, name_starts_with = NULL)
}
\arguments{
\item{name_like}{\itemize{
\item a string to match against part of the official
name field.  This can be a single string or character
vector of strings.
}}

\item{name_starts_with}{\itemize{
\item a string to match against the start of
the official name field. This can be a single string or
character vector of strings.
}}
}
\value{
a sf object containing the geometry or geometries of the
matching provincal parts.
}
\description{
Fetch Township objects from Geohub
}
\details{
This function will connect to the Geohub api and return the
matching records from the Ontario Township
endpoint.  The geohub details page can be found here:

\url{https://geohub.lio.gov.on.ca/datasets/lio::geographic-township-improved/about}

This functions currently supports filters for OFFICIAL_NAME like
or OFFICIAL_NAME starts with.  Please note that the geohub
api is case sensitive and that the Township names are stored in
uppercase in GeoHub.  This function will automatically convert any filter
arguments to upper case before sending the request.
}
\examples{

\donttest{
twnshp <- fetch_township(name_like='Saugeen')
unique(twnshp$OFFICIAL_NAME)

twnshp <- fetch_township(name_like=c('Saugeen', 'Bruce'))
unique(twnshp$OFFICIAL_NAME)

my_plot <- ggplot2::ggplot() +
 ggplot2::geom_sf(data = twnshp)
print(my_plot)
}
}
\author{
R. Adam Cottrill
}
