% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_waterbody.R
\name{fetch_waterbody}
\alias{fetch_waterbody}
\title{Fetch waterbody objects from Geohub}
\usage{
fetch_waterbody(
  wbylid = NULL,
  name_like = NULL,
  name_starts_with = NULL,
  name_ends_with = NULL
)
}
\arguments{
\item{wbylid}{\itemize{
\item a single waterbody id or a vector of waterbody
ids.
}}

\item{name_like}{\itemize{
\item a sting to match against part of the waterbody
name fields.
}}

\item{name_starts_with}{\itemize{
\item a sting to match against the start of
the waterbody name fields.
}}

\item{name_ends_with}{\itemize{
\item a sting to match against the start of the
waterbody name fields.
}}
}
\value{
sf (simple feature) object
}
\description{
Fetch waterbody objects from Geohub
}
\details{
This function will connect to the Geohub api and return the
matching records from the Ontario waterbody location identifier
endpoint.  The geohub details page can be fount here:
https://geohub.lio.gov.on.ca/datasets/lio::ontario-waterbody-location-identifier/about

This function current supports filters for waterbody body
identifiers, as well as partial string matches (names contains,
name starts with, and name ends with).  The string matches are
case sensitive, so 'Huron' will not match "Inverhuron".
}
\author{
R. Adam Cottrill
}
