% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_expressions.R
\name{starts_with_expression}
\alias{starts_with_expression}
\title{STARTS_WITH Query Expreession}
\usage{
starts_with_expression(field_name, arg)
}
\arguments{
\item{field_name}{\itemize{
\item the name of the field to query against
}}

\item{arg}{\itemize{
\item the value of the field to match
}}
}
\value{
formatted string
}
\description{
Build a 'starts with query string'
}
\details{
This is a helper function that accepts a field name and a value
and returns a formatted string of the form:
'\if{html}{\out{<fld>}}+LIKE+"\if{html}{\out{<value>}}\%\%"'.  This format is used by the ArcGis rest
api to return any matches were the field starts with the value.
This function is not intended to be used directly by users, but is
use by most of the other functions in OntarioGeoHubR
}
\author{
R. Adam Cottrill
}
